fix(db): 관리자 과제 목록 쿼리 PG 전용 ::int 캐스트 제거 — D1(SQLite) 구문 오류로 500 발생

+++ app/modules/submissions/submissions.server.ts
@@ -527,7 +527,7 @@ export const SubmissionHub = {
     const [list, counts] = await Promise.all([
       ctx.db.select().from(assignments).orderBy(desc(assignments.dueAt)),
       ctx.db
-        .select({ assignmentId: submissions.assignmentId, count: sql<number>`count(*)::int` })
+        .select({ assignmentId: submissions.assignmentId, count: sql<number>`count(*)` })
         .from(submissions)
         .groupBy(submissions.assignmentId),
     ]);
